// backend/src/application/services/authToken/authToken.service.interface.ts
import { RoleType, UserInterface } from '@/domain/entities/user/user.entity.interface';
import { UserTokenInterface } from '@/domain/entities/userToken/userToken.entity.interface';

export interface UserPayloadOptions {
  userId: string;
  email?: string;
  admin?: boolean;
  role?: RoleType;
}

export interface VerifyRefreshTokenOptions {
  token: UserTokenInterface | null;
  tokenValue: string;
  userId: string;
}

export interface AuthServiceInterface {
  generateAccessToken: (userId: string, user: UserInterface) => Promise<string>;
  generateRefreshToken: (userId: string) => UserTokenInterface;
  verifyAccessToken: (token: string, ignoreExpiration?: boolean) => Promise<UserPayloadOptions>;
  verifyRefreshToken: (options: VerifyRefreshTokenOptions) => UserTokenInterface;
}